:adhesive_bandage: service 구조 수정, Transactional annotation 붙임
